refactor(hashing): rework polynomial() hash-func as non-const
The primary reason for this rework is to introduce handling of
misalgined pointers - it's impossible to do it in a const-function,
so it was decided to drop this constraint.

Initially it was planned to make all hash functions to be const-only,
but now I came to conclusion that it's better to prioritize non-const
implementations first and then introduce const equivalents later for
specifically compile-time hashing.